Changed the file structure around a bit.
